Python Telegram chatbot/requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
requests>=2.31.0
pytz>=2023.3
openai>=1.0.0